refactoring: PollColumn 관련 컴포넌트 분리 및 파일 위치정리
PollColumn 컴포넌트를 단일 컴포넌트로 분리
PollColumn.js PollColumnTitle.js를 PollColumn 폴더로 이동
